% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVRMSE.R
\name{CVRMSE}
\alias{CVRMSE}
\title{Coefficient of Variation of Root Mean Squared Error}
\usage{
CVRMSE(actual, predicted)
}
\arguments{
\item{actual}{Actual value of the target variable}

\item{predicted}{Predicted/forecasted value of the target variable}
}
\value{
CVRMSE of the fitted model
}
\description{
Coefficient of Variation of Root Mean Squared Error
}
\examples{
actual <- c(100, 150, 200, 250, 300, 350, 400, 450, 500, 550)
predicted <- c(95, 148, 210, 245, 290, 360, 395, 440, 510, 540)
CVRMSE(actual, predicted)
}
