% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAE.R
\name{MAE}
\alias{MAE}
\title{Mean Absolute Error}
\usage{
MAE(actual, predicted)
}
\arguments{
\item{actual}{Actual value of the target variable}

\item{predicted}{Predicted/forecasted value of the target variable}
}
\value{
MAE of the fitted model
}
\description{
Mean Absolute Error
}
\examples{
 actual <- c(100, 150, 200, 250, 300, 350, 400, 450, 500, 550)
predicted <- c(95, 148, 210, 245, 290, 360, 395, 440, 510, 540)
MAE(actual, predicted)
}
