% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPE.R
\name{MAPE}
\alias{MAPE}
\title{Mean Absolute Percentage Error}
\usage{
MAPE(actual, predicted)
}
\arguments{
\item{actual}{Actual value of the target variable}

\item{predicted}{Predicted value of the target variable}
}
\value{
MAPE of the fitted model
}
\description{
Mean Absolute Percentage Error
}
\examples{
actual <- c(100, 150, 200, 250, 300, 350, 400, 450, 500, 550)
predicted <- c(95, 148, 210, 245, 290, 360, 395, 440, 510, 540)
MAPE(actual, predicted)
}
