% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME.R
\name{ME}
\alias{ME}
\title{Maximum Error}
\usage{
ME(actual, predicted)
}
\arguments{
\item{actual}{Actual value of the target variable}

\item{predicted}{Predicted value of the target variable}
}
\value{
ME of the fitted model
}
\description{
Maximum Error
}
\examples{
actual <- c(100, 150, 200, 250, 300, 350, 400, 450, 500, 550)
predicted <- c(95, 148, 210, 245, 290, 360, 395, 440, 510, 540)
ME(actual, predicted)
}
