% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRMSE.R
\name{NRMSE}
\alias{NRMSE}
\title{Normalised Root Mean Squared Error}
\usage{
NRMSE(actual, predicted)
}
\arguments{
\item{actual}{Actual value of the target variable}

\item{predicted}{Predicted value of the target variable}
}
\value{
NRMSE of the fitted model
}
\description{
Normalised Root Mean Squared Error
}
\examples{
actual <- c(100, 150, 200, 250, 300, 350, 400, 450, 500, 550)
predicted <- c(95, 148, 210, 245, 290, 360, 395, 440, 510, 540)
NRMSE(actual, predicted)
}
