% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSquare.R
\name{R2}
\alias{R2}
\title{Coefficient of Determination (R-Square)}
\usage{
R2(actual, predicted)
}
\arguments{
\item{actual}{Actual value of the target variable}

\item{predicted}{Predicted value of the target variable}
}
\value{
Coefficient of Determination (R-Square) of the fitted model
}
\description{
Coefficient of Determination (R-Square)
}
\examples{
actual <- c(100, 150, 200, 250, 300, 350, 400, 450, 500, 550)
predicted <- c(95, 148, 210, 245, 290, 360, 395, 440, 510, 540)
R2(actual, predicted)
}
